Move options processing to options.js
