+++ lib/loadtest.js
@@ -4,6 +4,7 @@ import * as httpClient from './httpClient.js'
 import * as websocket from './websocket.js'
 import {Latency} from './latency.js'
 import {HighResolutionTimer} from './hrtimer.js'
+import {processOptions} from './options.js'
 
 const SHOW_INTERVAL_MS = 5000;
 
@@ -32,28 +33,15 @@ https.globalAgent.maxSockets = 1000;
  * An optional callback will be called if/when the test finishes.
  */
 export function loadTest(options, callback) {
-	if (!options.url) {
-		console.error('Missing URL in options');
-		return;
-	}
-	options.concurrency = options.concurrency || 1;
-	if (options.requestsPerSecond) {
-		options.requestsPerSecond = options.requestsPerSecond / options.concurrency;
-	}
-	if (!options.url.startsWith('http://') && !options.url.startsWith('https://') && !options.url.startsWith('ws://')) {
-		console.error('Invalid URL %s, must be http://, https:// or ws://', options.url);
-		return;
-	}
-
-	if (options.url.startsWith('ws:')) {
-		if (options.requestsPerSecond) {
-			console.error('"requestsPerSecond" not supported for WebSockets');
+	processOptions(options, error => {
+		if (error) {
+			if (callback) return callback(error)
+			throw new error
 		}
-	}
-
-	const operation = new Operation(options, callback);
-	operation.start();
-	return operation;
+		const operation = new Operation(options, callback);
+		operation.start();
+		return operation;
+	})
 }
 
 /**
